% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellcyclefig2a-package.R
\docType{package}
\name{cellcyclefig2a-package}
\alias{cellcyclefig2a}
\alias{cellcyclefig2a-package}
\title{cellcyclefig2a: Reproduction de la Figure 2A de Kelliher et al. 2016}
\description{
Reproduit la Figure 2A de Kelliher et al. 2016 (PLOS Genetics) : une heatmap des genes periodiques du cycle cellulaire de Saccharomyces cerevisiae. Les profils d'expression sont normalises en z-score par gene, puis les genes sont ordonnes par phase du pic d'expression afin de faire apparaitre les vagues diagonales de transcription successives.
}
\author{
\strong{Maintainer}: Guillaume Gautreau \email{guillaume.gautreau44@gmail.com} (\href{https://orcid.org/0000-0002-0970-9361}{ORCID})

}
\keyword{internal}
