% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{heatmap_palette}
\alias{heatmap_palette}
\title{Palette de couleurs de la heatmap}
\usage{
heatmap_palette()
}
\value{
Un vecteur de trois couleurs hexadecimales.
}
\description{
Renvoie les trois couleurs de la colormap de l'article : cyan (expression
basse) vers noir (moyenne) vers jaune (expression haute).
}
\examples{
heatmap_palette()

}
